cleanup: merge redundant logic
Merge redundant modules after #972 0be1f0cc7b01 'SAM debug: show
CodeLenses in CFN templates'.
