+++ src/test/shared/sam/debugger/samDebugConfigProvider.test.ts
@@ -7,13 +7,7 @@ import * as assert from 'assert'
 import * as os from 'os'
 import * as path from 'path'
 import * as vscode from 'vscode'
-import {
-    AWS_SAM_DEBUG_TYPE,
-    CODE_TARGET_TYPE,
-    DIRECT_INVOKE_TYPE,
-    DotNetCoreDebugConfiguration,
-    TEMPLATE_TARGET_TYPE,
-} from '../../../../lambda/local/debugConfiguration'
+import { DotNetCoreDebugConfiguration } from '../../../../lambda/local/debugConfiguration'
 import * as lambdaModel from '../../../../lambda/models/samLambdaRuntime'
 import { CloudFormationTemplateRegistry } from '../../../../shared/cloudformation/templateRegistry'
 import { mkdir, rmrf } from '../../../../shared/filesystem'
@@ -21,11 +15,13 @@ import { makeTemporaryToolkitFolder } from '../../../../shared/filesystemUtiliti
 import {
     TemplateTargetProperties,
     AwsSamDebuggerConfiguration,
+    AWS_SAM_DEBUG_TYPE,
+    CODE_TARGET_TYPE,
+    DIRECT_INVOKE_TYPE,
+    TEMPLATE_TARGET_TYPE,
+    createAwsSamDebugConfig,
 } from '../../../../shared/sam/debugger/awsSamDebugConfiguration'
-import {
-    createDirectInvokeSamDebugConfiguration,
-    SamDebugConfigProvider,
-} from '../../../../shared/sam/debugger/awsSamDebugger'
+import { SamDebugConfigProvider } from '../../../../shared/sam/debugger/awsSamDebugger'
 import { SamLaunchRequestArgs } from '../../../../shared/sam/debugger/samDebugSession'
 import * as pathutil from '../../../../shared/utilities/pathUtils'
 import { FakeExtensionContext } from '../../../fakeExtensionContext'
@@ -588,12 +584,12 @@ describe('AwsSamDebugConfigurationProvider', async () => {
     })
 })
 
-describe('createDirectInvokeSamDebugConfiguration', () => {
+describe('createAwsSamDebugConfig', () => {
     const name = 'my body is a template'
     const templatePath = path.join('two', 'roads', 'diverged', 'in', 'a', 'yellow', 'wood')
 
     it('creates a template-type SAM debugger configuration with minimal configurations', () => {
-        const config = createDirectInvokeSamDebugConfiguration(name, templatePath)
+        const config = createAwsSamDebugConfig(name, templatePath)
         assert.strictEqual(config.invokeTarget.target, TEMPLATE_TARGET_TYPE)
         const invokeTarget = config.invokeTarget as TemplateTargetProperties
         assert.strictEqual(config.name, name)
@@ -612,7 +608,7 @@ describe('createDirectInvokeSamDebugConfiguration', () => {
             },
             dockerNetwork: 'rockerFretwork',
         }
-        const config = createDirectInvokeSamDebugConfiguration(name, templatePath, params)
+        const config = createAwsSamDebugConfig(name, templatePath, params)
         assert.deepStrictEqual(config.lambda?.event?.json, params.eventJson)
         assert.deepStrictEqual(config.lambda?.environmentVariables, params.environmentVariables)
         assert.strictEqual(config.sam?.dockerNetwork, params.dockerNetwork)
@@ -667,3 +663,35 @@ async function createAndRegisterYaml(
     await strToYamlFile(makeSampleSamTemplateYaml(true, subValues), file.fsPath)
     await registry.addTemplateToRegistry(file)
 }
+
+describe('createAwsSamDebugConfig', () => {
+    const name = 'my body is a template'
+    const templatePath = path.join('two', 'roads', 'diverged', 'in', 'a', 'yellow', 'wood')
+
+    it('creates a template-type SAM debugger configuration with minimal configurations', () => {
+        const config = createAwsSamDebugConfig(name, templatePath)
+        assert.strictEqual(config.invokeTarget.target, TEMPLATE_TARGET_TYPE)
+        const invokeTarget = config.invokeTarget as TemplateTargetProperties
+        assert.strictEqual(config.name, name)
+        assert.strictEqual(invokeTarget.samTemplateResource, name)
+        assert.strictEqual(invokeTarget.samTemplatePath, templatePath)
+        assert.ok(!config.hasOwnProperty('lambda'))
+    })
+
+    it('creates a template-type SAM debugger configuration with additional params', () => {
+        const params = {
+            eventJson: {
+                event: 'uneventufl',
+            },
+            environmentVariables: {
+                varial: 'invert to fakie',
+            },
+            dockerNetwork: 'rockerFretwork',
+        }
+        const config = createAwsSamDebugConfig(name, templatePath, params)
+        assert.deepStrictEqual(config.lambda?.event?.json, params.eventJson)
+        assert.deepStrictEqual(config.lambda?.environmentVariables, params.environmentVariables)
+        assert.strictEqual(config.sam?.dockerNetwork, params.dockerNetwork)
+        assert.strictEqual(config.sam?.containerBuild, undefined)
+    })
+})
